requirements.txt
flask
requests
beautifulsoup4
scikit-learn
tldextract
python-whois